backend/blog_backend/requirements.txt
asgiref==3.8.1
Django==5.0.11
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mssql-django==1.5
pillow==11.1.0
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
tzdata==2025.1
